[IMP] l10n_uy_edi: Add generic abstract model
Fixup conflicts 4 commits
Cherry pick of a6c104b9c54114aa4fff6599f7285880f418613d failed

stdout:
Auto-merging l10n_uy_edi/views/account_move_views.xml
Auto-merging l10n_uy_edi/models/account_move.py
CONFLICT (content): Merge conflict in l10n_uy_edi/models/account_move.py
Auto-merging l10n_uy_edi/models/__init__.py
Auto-merging l10n_uy_edi/__manifest__.py

stderr:
23:27:59.531072 git.c:439               trace: built-in: git cherry-pick a6c104b9c54114aa4fff6599f7285880f418613d
error: could not apply a6c104b... [ADD] l10n_uy_edi: Abstract model
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
----------
status:

 closes ingadhoc/uruguay#66

x-original-commit: 6d38bc400f583adda8f67396a442fc1ce7a69e84
x-original-commit: 5d260220c67f1eedfc9dccda34c9c1636b59e5ff
x-original-commit: i0f3888843ebb7852d5388dee0cf8bcd02d8dc2bb
x-original-commit: i776a8bb6407b30537ac2e9d7ed4f888c37cd264c
x-original-commit: iia83b6ec8801dd8b42dd8c443b357b278c541768c
