+++ l10n_uy_edi/__manifest__.py
@@ -7,7 +7,7 @@
     'author': 'ADHOC SA',
     'category': 'Localization',
     'license': 'LGPL-3',
-    'version': '13.0.1.10.0',
+    'version': '15.0.1.1.0',
     'depends': [
         'l10n_uy_account',
         'account_debit_note',
