+++ l10n_uy_edi/models/__init__.py
@@ -1,6 +1,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import res_company
 from . import res_config_settings
+from . import l10n_uy_cfe
 from . import account_move
 from . import account_journal
 from . import account_tax
