+++ l10n_uy_edi/views/account_move_views.xml
@@ -39,7 +39,7 @@
                                         <field name="l10n_uy_ucfe_notif"/>
                                         <field name="l10n_uy_cfe_partner_status"/>
                                         <field name="l10n_uy_cfe_uuid"/>
-                                        <button name="action_l10n_uy_get_uruware_inv" type="object" string="Get Uruware Invoice" class="oe_inline oe_link"/>
+                                        <button name="action_l10n_uy_get_uruware_cfe" type="object" string="Get Uruware Invoice" class="oe_inline oe_link"/>
                                         <field name="l10n_uy_cfe_file"/>
                                     </group>
                                     <group string="DGI">
@@ -48,7 +48,6 @@
                                         <div>
                                             <field name="l10n_uy_cfe_dgi_state"/>
                                         </div>
-                                        <field name="l10n_uy_dgi_acceptation_status"/>
                                         <field name="l10n_uy_dgi_barcode"/>
                                     </group>
                                 </page>
